+++ src/MusicalScore/Graphical/MusicSheetCalculator.ts
@@ -2187,9 +2187,11 @@ export abstract class MusicSheetCalculator {
                                     if (note === note.NoteTie.Notes.last()) {
                                         continue; // nothing to do on last note. don't create last tie twice.
                                     }
-                                    for (const gTie of startStaffEntry.GraphicalTies) {
-                                        if (gTie.Tie === tie) {
-                                            continue; // don't handle the same tie on the same startStaffEntry twice
+                                    if (startStaffEntry) {
+                                        for (const gTie of startStaffEntry.GraphicalTies) {
+                                            if (gTie.Tie === tie) {
+                                                continue; // don't handle the same tie on the same startStaffEntry twice
+                                            }
                                         }
                                     }
                                     this.handleTie(tie, startStaffEntry, staffIndex, measureIndex);
@@ -2583,12 +2585,8 @@ export abstract class MusicSheetCalculator {
         }
     }
 
-    private calculateSkyBottomLines(): void {
-        for (const musicSystem of this.musicSystems) {
-            for (const staffLine of musicSystem.StaffLines) {
-                staffLine.SkyBottomLineCalculator.calculateLines();
-            }
-        }
+    protected calculateSkyBottomLines(): void {
+        // override
     }
 
     /**